requirements.txt
gradio
python-dotenv
openai
markitdown[all]